controllers Reservation and Review created

+++ src/main/java/cristiancicale/capstone/services/ReservationService.java
@@ -3,6 +3,7 @@ package cristiancicale.capstone.services;
 import cristiancicale.capstone.entities.Event;
 import cristiancicale.capstone.entities.Reservation;
 import cristiancicale.capstone.entities.User;
+import cristiancicale.capstone.exceptions.AccessDeniedException;
 import cristiancicale.capstone.exceptions.BadRequestException;
 import cristiancicale.capstone.exceptions.NotFoundException;
 import cristiancicale.capstone.payloads.ReservationDTO;
@@ -47,19 +48,40 @@ public class ReservationService {
         return reservationRepository.save(reservation);
     }
 
-    public Page<Reservation> findAll(int page, int size, String sortBy) {
+    public Page<Reservation> findAll(int page, int size) {
         if (size > 10 || size < 0) size = 10;
         if (page < 10) page = 0;
-        Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy));
+        Pageable pageable = PageRequest.of(page, size, Sort.by("createdAt").descending());
         return this.reservationRepository.findAll(pageable);
     }
 
-    public Reservation findById(UUID id) {
-        return reservationRepository.findById(id).orElseThrow(() -> new NotFoundException(id));
+    public Reservation findById(UUID id, User currentUser) {
+
+        Reservation found = reservationRepository.findById(id).orElseThrow(() -> new NotFoundException(id));
+
+        boolean isOwner = found.getUser().getId().equals(currentUser.getId());
+
+        boolean isAdmin = currentUser.getRole().name().equals("ROLE_ADMIN");
+
+        if (!isOwner && !isAdmin) {
+            throw new AccessDeniedException("Non puoi accedere a questa prenotazione");
+        }
+
+        return found;
     }
 
-    public void findByIdAndDelete(UUID id) {
-        Reservation found = findById(id);
+    public void findByIdAndDelete(UUID id, User currentUser) {
+
+        Reservation found = reservationRepository.findById(id).orElseThrow(() -> new NotFoundException(id));
+
+        boolean isOwner = found.getUser().getId().equals(currentUser.getId());
+
+        boolean isAdmin = currentUser.getRole().name().equals("ROLE_ADMIN");
+
+        if (!isOwner && !isAdmin) {
+            throw new AccessDeniedException("Non puoi eliminare questa prenotazione");
+        }
+
         reservationRepository.delete(found);
     }
 }
